requirements.txt
jupyter
streamlit==1.40.0
google-generativeai
python-dotenv